fixed issues with pull request
