+++ arc/__init__.py
@@ -1,3 +1,4 @@
+from .augmentations import *
 from .agents import Agent
 from .eval import evaluate_and_report
 from .hints import BoardHints, Hints
