--- arc/augmentations/augmentations/functional/__init__.py
@@ -1,3 +1,0 @@
-from augmentations.functional.spatial import *
-from augmentations.functional.color import *
-from augmentations.functional.noise import *
