+++ arc/augmentations/functional/__init__.py
@@ -1,0 +1,3 @@
+from arc.augmentations.functional.spatial import *
+from arc.augmentations.functional.color import *
+from arc.augmentations.functional.noise import *
